Twitter/requirements.txt
Django==2.2.12
django-cors-headers==3.7.0
djangorestframework==3.12.4
dj-rest-auth==2.1.9
djangorestframework_simplejwt==4.6.0
django-allauth==0.43.0

